requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
tensorflow==2.13.0
pillow==9.4.0
numpy==1.24.3
python-multipart
